int to double

+++ phase1/src/main/java/atm/ChequingAccount.java
@@ -14,13 +14,13 @@ public class ChequingAccount extends Account {
         return "Chequing Account";
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return this.balance;
     }
 
     // this takes in either a negative balance if money was taken from the account
     // or positive balance if money is added
-    public void setBalance(int amount) {
+    public void setBalance(double amount) {
 
         if  (this.balance < 100)
